Updated the build process and ran itis with new fixed function
I updated the species list build process to output JSON documents from the xDD search results off the literature list and a JSON output for the species list as opposed to a simple text list of the names. I also ran and produced the ITIS output for the species list.